+++ front-end-m.c-plastico/src/components/menu-lateral/menu.css
@@ -9,8 +9,17 @@
   left: 0;
   border-top-right-radius: 20px;
   border-bottom-right-radius: 20px;
-   border: 2px solid black;
+  z-index: 9999;
+}
 
+.sidebar-overlay {
+  position: fixed;
+  top: 0;
+  left: 0;
+  width: 100vw;
+  height: 100vh;
+  background-color: rgba(0, 0, 0, 0.5);
+  z-index: 999;
 }
 
 .sidebar-title {
