+++ front-end-m.c-plastico/src/components/popup-check/popup-check.css
@@ -8,7 +8,7 @@
   display: flex;
   justify-content: center;
   align-items: center;
-  z-index: 9999;
+  z-index: 9998;
   animation: fadeIn 0.3s ease;
 }
