--- front-end-m.c-plastico/src/pages/menu_lateral/index.js
@@ -1,10 +1,0 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
-import AppView from './menu-lateral/view/menu';
-
-ReactDOM.render(
-  <React.StrictMode>
-    <AppView />
-  </React.StrictMode>,
-  document.getElementById('root')
-);
